fix(autoware_kalman_filter): fix bugprone-narrowing-conversions warnings (#930)

+++ common/autoware_kalman_filter/src/time_delay_kalman_filter.cpp
@@ -23,7 +23,7 @@ void TimeDelayKalmanFilter::init(
   const Eigen::MatrixXd & x, const Eigen::MatrixXd & P0, const int max_delay_step)
 {
   max_delay_step_ = max_delay_step;
-  dim_x_ = x.rows();
+  dim_x_ = static_cast<int>(x.rows());
   dim_x_ex_ = dim_x_ * max_delay_step;
 
   x_ = Eigen::MatrixXd::Zero(dim_x_ex_, 1);
@@ -90,7 +90,7 @@ bool TimeDelayKalmanFilter::updateWithDelay(
     return false;
   }
 
-  const int dim_y = y.rows();
+  const int dim_y = static_cast<int>(y.rows());
 
   /* set measurement matrix */
   Eigen::MatrixXd C_ex = Eigen::MatrixXd::Zero(dim_y, dim_x_ex_);
